add CoronaNewsCrawler
