+++ app/crawler/utils/news.py
@@ -1,0 +1,10 @@
+import asyncio
+import aiohttp
+
+from bs4 import BeautifulSoup
+from ...ext import Performance
+
+
+class CoronaNewsCrawler:
+    def __init__(self):
+        self.loop = Performance()
